Implement ending ongoing entries with

+++ main.py
@@ -62,7 +62,7 @@ if __name__ == "__main__":
 			description = args[0]
 			file_path   = args[1]
 
-		parser = Parser(debug=True)
+		parser = Parser(debug=False)
 		with open(file_path, 'r') as file:
 			tree = parser.parse(file.read())
 
@@ -76,7 +76,7 @@ if __name__ == "__main__":
 		description = args[0]
 		file_path   = args[1]
 
-		parser = Parser(debug=True)
+		parser = Parser(debug=False)
 		with open(file_path, 'r') as file:
 			tree = parser.parse(file.read())
 
@@ -150,6 +150,54 @@ if __name__ == "__main__":
 		, ongoing=True)
 
 		print(tree.format())
+
+		with open(file_path, 'w') as file:
+			file.write(tree.format())
+
+	if cmd == 'end':
+		if len(args) < 2:
+			print('Error: expected an argument')
+			exit(1)
+
+		description = args[0]
+		file_path   = args[1]
+
+		parser = Parser(debug=False)
+		with open(file_path, 'r') as file:
+			tree = parser.parse(file.read())
+
+		finds = tree.find(description, ongoing = True, with_parent = True)
+
+		if len(finds) == 0:
+			print(f'No ongoing entry with the name "{description}"')
+			exit(0)
+		elif len(finds) > 1:
+			print(f'There are multiple ongoing entries with the name "{description}":')
+			print()
+			print(''.join(['\t' + f[0].format(full_time = True) for f in finds]))
+			while True:
+				print(f'Which one would you like to end?')
+				index = input(f'[1-{len(finds)}]: ')
+				if index.isnumeric() and int(index) > 0 and int(index) <= len(finds):
+					segment, parent = finds[int(index)-1]
+					break
+				else:
+					if index == 'none':
+						exit(0)
+					print(f'Invalid index "{index}"')
+					print(f'(type "none" to cancle)')
+		else:
+			print(f'Currently ongoing entry "{description}":')
+			print()
+			print(''.join(['\t' + f[0].format(full_time = True) for f in finds]))
+			print('Are you sure you want to end this entry?')
+			ans = input('(y/n): ')
+			if ans != 'y':
+				exit(0)
+
+			segment, parent = finds[0]
+
+		parent.segments = [e for e in parent.segments if e is not segment]
 
 		with open(file_path, 'w') as file:
 			file.write(tree.format())
